requirements.txt
Django 
djangorestframework 
celery[redis]
psycopg2-binary
requests